IMPALA-7551
The call stack here is impala-server-hs3 ExecuteStatement ->
client-request-state Wait / WaitAsync (this is what actually prints Rows
Available) and then WaitInternal which calls GetCoordinator --> Wait
which calls fragment-instance-state WaitForOpen which blocks on a
promise which is notified when FragmentInstanceState calls Open which I
think means once the nodes in the fragment are all opened

Think this is the right approach. PlanRootSink::GetNext is only called
when a client issues a Thrift request for FetchResults

PlanRootSink::Send is triggered whenever the fragment instance has
produced rows.

Change-Id: I315722c397e96404c9503d692b8712e172133f3d