[WIP] IMPALA-8656: Support for eagerly fetching and spooling all query result rows
The current implementation of PlanRootSink has several issues, which are
documented in IMPALA-8656 and IMPALA-4268. To summarize the current
design, FragmentInstanceState will repeatedly call GetNext on the ExecNode
tree and send RowBatches to a DataSink using the DataSink::Send method. For
INSERT queries the DataSink can be a HdfsTableSink or KuduTableSink. For
SELECT queries the DataSink is called the PlanRootSink. The PlanRootSink
will buffer a single RowBatch sent to it by the FragmentInstanceState,
and then will block the FragmentInstanceState from sending any more
RowBatches until the client thread fetches the buffered RowBatch. The
client thread is created by the ImpalaServer, which ultimately makes a
call to PlanRootSink::GetNext using the classes ClientRequestState and
Coordinator.

This design allows for back-pressure to be introduced to the fragment
execution. However, this design also has several drawbacks. The main
one this patch addresses is that fragment resources are kept alive until
the client consumes all rows, which can happen frequently in
applications such as Hue.

This patch re-works the PlanRootSink so that it writes RowBatches to a
BufferedTupleStream rather than buffering a single RowBatch at a time.
The goal is to buffer all results in the BufferedTupleStream and then
release all fragment level resources. The client thread then reads from
the BufferedTupleStream via PlanRootSink::GetNext.

The implementation of this patch requires changes to parts of the
the fe/ and be/ code.

be/ changes:
* PlanRootSink creates and manages a BufferedTupleStream as well as a
ReservationManager for admission control integration
* The BufferedTupleStream is pinned in memory, consideration for
spill-to-disk will be added in a future patch; however, if the BTS is
full the old behavior should kick in (e.g. back-pressure)
* New methods have been added to PlanRootSink to release the BTS and
ReservationManager once all results have been fetched from the query
* The ReservationManager is linked to the query-level resource pool so
that the BufferedTupleStream resources can live on after the fragment is
closed
* The previous implementation of PlanRootSink, ran the
ScalarExprEvaluators in the client thread; however, this is no longer
possible in the new design because the fragment-level codegen
resources are closed before rows are necessarily fetched. To fix this,
Tuple::MaterializeExprs is called in PlanRootSink::Send; this requires
correspoding changes to QueryResultSet as well.

fe/ changes:
* ResourceProfile integration for the PlanRootSink - this is still a WIP
and will require some re-factoring because currently DataSink
ResourceProfiles are accounted in the fragment-level pool whereas the
PlanRootSink profile needs to be accounted for at the query level
* DataSinks currently do not have a RowDescriptor, instead they take the
RowDescriptor of the top-most node in the exec_tree_; this is
problematic for the new PlanRootSink because it has to materialize all
ScalarExprs and write them to a new RowBatch. The fix is to change the
fe/ so it creates the correct RowDescriptor for the PlanRootSink; this
requires changes to some of the Thrift files as well.

Implementation TODO:
* Looks like there is a bug where only one row batch is returned ever
* Figure out ResourceProfile changes in fe
* Make changes for HS2ResultSet
* Double check deletion of any DCHECKS
* General code cleanup and re-formatting
* Javadocs

Updates:
* Initial work to get admission control to work properly in fe/
* Added support for fetching HS2 ResultSets
* Updated sychronization and cancellation logic in PlanRootSink

Various bug fixes

* nullptr check added to PlanRootSink::Close
* Proper cancellation handling in PlanRootSink::Cancel
* NULL printing fix in ASCII QueryResultSet
* be/ test fixes in test-env.cc

Change-Id: I0d6f4c53ff00bd70ea2cfa06d1d677d085dfcb1c