IMPALA-8656: Re-factor PlanRootSink into blocking and buffered implementations
Refactors PlanRootSink into a base class with two subclasses:
BlockingPlanRootSink and BufferedPlanRootSink. BlockingPlanRootSink
encapsulates the current implementation of PlanRootSink and
BufferedPlanRootSink encapsulates a new implementation that will buffer
RowBatches in memory until they are read by the client. The
implementation of BlockingPlanRootSink is left to future work.

A new query option called POOL_QUERY_RESULTS controls whether a
BlockingPlanRootSink or a BufferedPlanRootSink is used as the DataSink.
POOL_QUERY_RESULTS is false by default.

Added a few more docs to PlanRootSink and BlockingPlanRootSink to make
the implementation easier to understand.

Testing:
* Since no new functionality has been added no tests were added
* Running core tests

Change-Id: I8786b1a9af68ab0a8a094970d8f955eb20d04bca