IMPALA-9199 [POC][WIP]: Add support for single query retries on cluster membership changes
Adds the core logic for transparently retrying queries that fail due to
cluster membership changes.

The design of this feature is described in the JIRA, but the TL;DR is
that whenever a ClientRequestState receives an error Status for a query,
it checks if the Status is "Retryable". If it is, then it schedules a
retry of the query using a "query retry" threadpool.

This feature requires touching several parts of the query lifecycle /
Coordinator code:

* The feature is configurable ('retry_failed_queries') and is off by
default.

* Changes the ClientRequestState so that it can take in an existing
TExecRequest (this is required when retrying queries because the
TExecRequest of the failed query is copied and used for the
ClientRequestState of the retried query).

* ClientRequestState::ExecState is extended with three new states:
RETRYING, RETRIED, and UNKNOWN.

* ImpalaServer::RetryQueryFromThreadPool implements the core logic to
actually retry a failed query.

* Retries are transparent to the user. This is achieved by registering the
query id of the failed query with the ClientRequestState of the retried
query. This requires modifying the ImpalaServer
client_request_state_map_ so that different query ids can correspond to
the same ClientRequestState.

Change-Id: I2e4a0e72a9bf8ec10b91639aefd81bef17886ddd