IMPALA-9229: impala-shell 'profile' to show original and retried queries
All tests passing, needs a lot of cleanup.

TODO:
* Address all the hacks in GetRuntimeProfile in ImpalaServer and
  ImpalaHS2Server
* Specifically GetRuntimeProfile needs a version to get the exact
  query requested + a version to get the retried + original query
  regardless of which query attempt id is specified

One proposal for doing this in making parameters profile and
original_profile; profile is always either the exact profile requested
or the most recent attempt, depenfing on if original_profile is nullptr
or not. Actually might be cleaned to just include a boolean flag called
'include_original_profile' that controls this behavior.

Change-Id: I89cee02947b311e7bf9c7274f47dfc7214c1bb65